+++ src/__generated__/QueryRenderersRegistrationFlowQuery.graphql.ts
@@ -1,0 +1,217 @@
+/* tslint:disable */
+
+import { ConcreteRequest } from "relay-runtime";
+export type QueryRenderersRegistrationFlowQueryVariables = {
+    readonly saleID: string;
+};
+export type QueryRenderersRegistrationFlowQueryResponse = {
+    readonly sale: ({
+        readonly name: string | null;
+    }) | null;
+    readonly me: ({
+    }) | null;
+};
+
+
+
+/*
+query QueryRenderersRegistrationFlowQuery(
+  $saleID: String!
+) {
+  sale(id: $saleID) {
+    name
+    ...RegistrationFlow_sale
+    __id
+  }
+  me {
+    ...RegistrationFlow_me
+    __id
+  }
+}
+
+fragment RegistrationFlow_sale on Sale {
+  ...Registration_sale
+  __id
+}
+
+fragment RegistrationFlow_me on Me {
+  ...Registration_me
+  __id
+}
+
+fragment Registration_me on Me {
+  has_credit_cards
+  __id
+}
+
+fragment Registration_sale on Sale {
+  id
+  end_at
+  is_preview
+  live_start_at
+  name
+  start_at
+  __id
+}
+*/
+
+const node: ConcreteRequest = (function(){
+var v0 = [
+  {
+    "kind": "LocalArgument",
+    "name": "saleID",
+    "type": "String!",
+    "defaultValue": null
+  }
+],
+v1 = [
+  {
+    "kind": "Variable",
+    "name": "id",
+    "variableName": "saleID",
+    "type": "String!"
+  }
+],
+v2 = {
+  "kind": "ScalarField",
+  "alias": null,
+  "name": "name",
+  "args": null,
+  "storageKey": null
+},
+v3 = {
+  "kind": "ScalarField",
+  "alias": null,
+  "name": "__id",
+  "args": null,
+  "storageKey": null
+};
+return {
+  "kind": "Request",
+  "operationKind": "query",
+  "name": "QueryRenderersRegistrationFlowQuery",
+  "id": "eae7174e3cddc451acab97c6294292ce",
+  "text": null,
+  "metadata": {},
+  "fragment": {
+    "kind": "Fragment",
+    "name": "QueryRenderersRegistrationFlowQuery",
+    "type": "Query",
+    "metadata": null,
+    "argumentDefinitions": v0,
+    "selections": [
+      {
+        "kind": "LinkedField",
+        "alias": null,
+        "name": "sale",
+        "storageKey": null,
+        "args": v1,
+        "concreteType": "Sale",
+        "plural": false,
+        "selections": [
+          v2,
+          {
+            "kind": "FragmentSpread",
+            "name": "RegistrationFlow_sale",
+            "args": null
+          },
+          v3
+        ]
+      },
+      {
+        "kind": "LinkedField",
+        "alias": null,
+        "name": "me",
+        "storageKey": null,
+        "args": null,
+        "concreteType": "Me",
+        "plural": false,
+        "selections": [
+          {
+            "kind": "FragmentSpread",
+            "name": "RegistrationFlow_me",
+            "args": null
+          },
+          v3
+        ]
+      }
+    ]
+  },
+  "operation": {
+    "kind": "Operation",
+    "name": "QueryRenderersRegistrationFlowQuery",
+    "argumentDefinitions": v0,
+    "selections": [
+      {
+        "kind": "LinkedField",
+        "alias": null,
+        "name": "sale",
+        "storageKey": null,
+        "args": v1,
+        "concreteType": "Sale",
+        "plural": false,
+        "selections": [
+          v2,
+          {
+            "kind": "ScalarField",
+            "alias": null,
+            "name": "id",
+            "args": null,
+            "storageKey": null
+          },
+          {
+            "kind": "ScalarField",
+            "alias": null,
+            "name": "end_at",
+            "args": null,
+            "storageKey": null
+          },
+          {
+            "kind": "ScalarField",
+            "alias": null,
+            "name": "is_preview",
+            "args": null,
+            "storageKey": null
+          },
+          {
+            "kind": "ScalarField",
+            "alias": null,
+            "name": "live_start_at",
+            "args": null,
+            "storageKey": null
+          },
+          {
+            "kind": "ScalarField",
+            "alias": null,
+            "name": "start_at",
+            "args": null,
+            "storageKey": null
+          },
+          v3
+        ]
+      },
+      {
+        "kind": "LinkedField",
+        "alias": null,
+        "name": "me",
+        "storageKey": null,
+        "args": null,
+        "concreteType": "Me",
+        "plural": false,
+        "selections": [
+          {
+            "kind": "ScalarField",
+            "alias": null,
+            "name": "has_credit_cards",
+            "args": null,
+            "storageKey": null
+          },
+          v3
+        ]
+      }
+    ]
+  }
+};
+})();
+(node as any).hash = 'bce11602f08e3a5b15eae2a068725187';
+export default node;