+++ src/lib/Containers/__tests__/index-tests.tsx
@@ -10,6 +10,7 @@ it("should export all components", () => {
     "Home",
     "Gene",
     "Sale",
+    "RegistrationFlow",
     "WorksForYou",
     "MyProfile",
     "Inbox",